project2/pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.98.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
